requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.6.15
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==42.0.0
daphne==4.2.0
Django==5.2.3
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.12.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
kafka-python==2.0.2
msgpack==1.1.1
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==25.1.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.0
service-identity==24.2.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
zope.interface==7.2
